+++ packages/paraglide/paraglide-js-adapter-sveltekit/example/src/params/lang.ts
@@ -1,0 +1,2 @@
+import { isAvailableLanguageTag } from "$paraglide/runtime"
+export const match = isAvailableLanguageTag