+++ packages/paraglide/paraglide-js/src/cli/commands/compile.ts
@@ -1,61 +1,90 @@
 import { loadProject, type InlangProject } from "@inlang/sdk"
-import consola from "consola"
 import { compile } from "../../compiler/compile.js"
 import fs from "node:fs/promises"
 import { resolve } from "node:path"
 import { Command } from "commander"
+import { telemetry } from "../../services/telemetry/implementation.js"
+import { writeOutput } from "../../services/file-handling/write-output.js"
+import { Logger } from "../../services/logger/index.js"
 
 export const compileCommand = new Command()
 	.name("compile")
 	.summary("Compiles inlang Paraglide-JS.")
-	.requiredOption("--project <path>", "The path to the inlang project.")
-	.requiredOption("--outdir <path>", "The path to the output directory.", "./src/paraglide")
-	.action(async (options: { project: string; outdir: string }) => {
-		consola.info(`Compiling inlang project at "${options.project}".`)
-
+	.requiredOption("--project <path>", 'The path to the inlang project. Example: "./project.inlang"')
+	.requiredOption(
+		"--outdir <path>",
+		'The path to the output directory. Example: "./src/paraglide"',
+		"./src/paraglide"
+	)
+	.requiredOption("--watch", "Watch for changes and recompile.", false)
+	.action(async (options: { project: string; outdir: string; watch: boolean }) => {
+		const logger = new Logger({ silent: false, prefix: true })
 		const path = resolve(process.cwd(), options.project)
-
-		const project = exitIfErrors(
-			await loadProject({
-				settingsFilePath: path,
-				nodeishFs: fs,
-			})
-		)
-
-		const output = compile({
-			messages: project.query.messages.getAll(),
-			settings: project.settings(),
-		})
-
 		const outputDirectory = resolve(process.cwd(), options.outdir)
 
-		// create the compiled-output directory if it doesn't exist
-		await fs.access(outputDirectory).catch(async () => {
-			await fs.mkdir(outputDirectory, { recursive: true })
-		})
+		logger.info(`Compiling inlang project at "${options.project}".`)
+		const project = exitIfErrors(
+			await loadProject({
+				projectPath: path,
+				nodeishFs: fs,
+				_capture(id, props) {
+					telemetry.capture({
+						// @ts-ignore the event types
+						event: id,
+						properties: props,
+					})
+				},
+			}),
+			logger
+		)
 
-		// create the messages directory if it doesn't exist
-		await fs.access(outputDirectory + "/messages").catch(async () => {
-			await fs.mkdir(outputDirectory + "/messages")
-		})
-
-		for (const [fileName, fileContent] of Object.entries(output)) {
-			await fs.writeFile(`${outputDirectory}/${fileName}`, fileContent, {
-				encoding: "utf-8",
+		async function execute() {
+			const output = compile({
+				messages: project.query.messages.getAll(),
+				settings: project.settings(),
 			})
+
+			await writeOutput(outputDirectory, output, fs)
 		}
 
-		consola.success("Successfully compiled the project.")
+		await execute()
+
+		// await Promise that never resolves to keep the process alive
+		if (options.watch) {
+			process.on("SIGINT", () => {
+				//start with a new line, since the ^C is on the current line
+				logger.ln().info("Stopping the watcher.")
+				process.exit(0)
+			})
+
+			let numChanges = 0
+			project.query.messages.getAll.subscribe(async (messages) => {
+				numChanges++
+				if (messages.length === 0) return //messages probably haven't loaded yet
+				if (numChanges === 1) return //don't recompile on the first run
+
+				logger.info("Messages changed. Recompiling...")
+				await execute()
+			})
+
+			/* eslint-disable no-constant-condition */
+			while (true) {
+				// Keep the process alive
+				await new Promise((resolve) => setTimeout(resolve, 10_000))
+			}
+		}
+
+		logger.info("Sucessfully compiled the project.")
 	})
 
 /**
  * Utility function to exit when the project has errors.
  */
-const exitIfErrors = (project: InlangProject) => {
+const exitIfErrors = (project: InlangProject, logger: Logger) => {
 	if (project.errors().length > 0) {
-		consola.warn(`The project has errors:`)
+		logger.warn(`The project has errors:`)
 		for (const error of project.errors()) {
-			consola.error(error)
+			logger.error(error)
 		}
 		process.exit(1)
 	}