+++ packages/sdk/src/createNodeishFsWithAbsolutePaths.test.ts
@@ -2,16 +2,16 @@ import { it, expect, vi } from "vitest"
 import { createNodeishFsWithAbsolutePaths } from "./createNodeishFsWithAbsolutePaths.js"
 import type { NodeishFilesystemSubset } from "./versionedInterfaces.js"
 
-it("throws an error if settingsFilePath is not an absolute path", () => {
+it("throws an error if projectPath is not an absolute path", () => {
 	const relativePath = "relative/path"
 
 	expect(() =>
-		createNodeishFsWithAbsolutePaths({ settingsFilePath: relativePath, nodeishFs: {} as any })
+		createNodeishFsWithAbsolutePaths({ projectPath: relativePath, nodeishFs: {} as any })
 	).toThrow()
 })
 
 it("intercepts paths correctly for readFile", async () => {
-	const settingsFilePath = `/Users/samuel/Documents/paraglide/example/project.inlang.json`
+	const projectPath = `/Users/samuel/Documents/paraglide/example/project.inlang`
 
 	const filePaths = [
 		["file.txt", `/Users/samuel/Documents/paraglide/example/file.txt`],
@@ -32,7 +32,7 @@ it("intercepts paths correctly for readFile", async () => {
 	} satisfies Record<keyof NodeishFilesystemSubset, any>
 
 	const interceptedFs = createNodeishFsWithAbsolutePaths({
-		settingsFilePath,
+		projectPath,
 		nodeishFs: mockNodeishFs,
 	})