+++ packages/website/src/pages/index/Footer.tsx
@@ -9,27 +9,6 @@ import { LanguagePicker } from "./LanguagePicker.jsx"
 import Link from "#src/renderer/Link.jsx"
 
 const Footer = () => {
-	const getProductsLinks = () => {
-		return [
-			{
-				name: m.marketplace_header_apps_title(),
-				href: "/c/apps",
-			},
-			{
-				name: m.marketplace_header_libraries_title(),
-				href: "/c/libraries",
-			},
-			{
-				name: m.marketplace_header_plugins_title(),
-				href: "/c/plugins",
-			},
-			{
-				name: m.marketplace_header_lintRules_title(),
-				href: "/c/lint-rules",
-			},
-		]
-	}
-
 	const socialMediaLinks = [
 		{
 			name: "X",
@@ -54,16 +33,21 @@ const Footer = () => {
 	const getResourceLinks = () => {
 		return [
 			{
+				name: m.footer_resources_documentation(),
+				href: "/documentation",
+			},
+			{
 				name: m.footer_resources_roadmap(),
 				href: "https://github.com/orgs/inlang/projects?query=is%3Aopen",
 			},
 			{
-				name: m.footer_documentation_title(),
-				href: "/documentation",
+				name: m.footer_resources_guides(),
+				href: "/c/guides",
 			},
 		]
 	}
-	const getContactLinks = () => {
+
+	const getMoreLinks = () => {
 		return [
 			{
 				name: m.footer_contact_getInTouch(),
@@ -78,6 +62,29 @@ const Footer = () => {
 				href: "https://github.com/inlang/monorepo/discussions/categories/feedback",
 			},
 			{ name: m.footer_contact_blog(), href: "/blog" },
+			{
+				name: m.footer_privacy_policy(),
+				href: "https://www.iubenda.com/privacy-policy/63446420",
+			},
+		]
+	}
+
+	const getCategoriesLinks = () => {
+		return [
+			{
+				name: m.footer_categories_apps(),
+				href: "/c/apps",
+			},
+			{
+				name: m.footer_categories_libraries(),
+				href: "/c/libraries",
+			},
+			{
+				name: m.footer_categories_plugins(),
+				href: "/c/plugins",
+			},
+			{ name: m.footer_categories_lintrules(), href: "/c/lint-rules" },
+			{ name: m.footer_categories_lix(), href: "/c/lix" },
 		]
 	}
 
@@ -118,8 +125,8 @@ const Footer = () => {
 					</For>
 				</div>
 				<div class="w-full sm:w-1/3 md:w-1/4 xl:px-4 flex flex-col pt-2">
-					<p class="font-semibold text-surface-900 pb-3">{m.footer_category_title()}</p>
-					<For each={getProductsLinks()}>
+					<p class="font-semibold text-surface-900 pb-3">Categories</p>
+					<For each={getCategoriesLinks()}>
 						{(link) => (
 							<div class="w-fit opacity-80">
 								<Button type="text" href={link.href}>
@@ -131,7 +138,7 @@ const Footer = () => {
 				</div>
 				<div class="w-full sm:w-1/3 md:w-1/4 xl:px-4 xl:flex flex-col pt-2">
 					<p class="font-semibold text-surface-900 pb-3">{m.footer_contact_title()}</p>
-					<For each={getContactLinks()}>
+					<For each={getMoreLinks()}>
 						{(link) => (
 							<div class="w-fit opacity-80">
 								<Button type="text" href={link.href}>